requirements.txt
glfw==2.3.0
PyOpenGL==3.1.5
pyrr==0.10.3
pillow==8.4.0
keyboard==0.13.5
